requirements.txt
flask
python-dotenv
PyMuPDF
langchain
langchain-core
langchain-community
langchain-groq
langchain-google-genai
langchain-text-splitters
faiss-cpu
tiktoken
